fix: address React Doctor findings
