+++ libs/ui/package.json
@@ -17,7 +17,7 @@
       "./src/chess/*.ts"
     ],
     "./layout/*": "./src/layout/*.tsx",
-    "./icons": "./src/icons/index.tsx",
+    "./icons": "./src/icons/index.ts",
     "./icons/*": "./src/icons/*.tsx",
     "./charts/*": "./src/charts/*.tsx"
   },
